moves journal formatting into multiple classes
To better allow for customization, formatting journal entries is moved
to separate classes. One can register for each field of a journaled
object the same way like before (e.g. register_on_journal_formatter
(:plaintext, 'subject')). The difference is purely inside the
JournalFormatter. It determines the formatter required for the given
field and class, instantiates it and calls the render method on the
instance.

JournalFormatter::Base subclasses are provided to handle the most common
types of formatting (e.g. plaintext and decimal).

The OpenProject formatters are removed altogether from
acts_as_journalized and are moved into the OpenProject /lib directory.

The helper_issues_show_detail_after_setting hook is removed. It should
no longer be needed.